requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
python-dotenv>=0.21.0
pandas>=2.0.0
python-dateutil>=2.8.2
typer>=0.9.0
rich>=13.0.0
questionary>=2.0.1
pytest>=8.0.0
pytest-cov>=6.0.0
